move analytic_beam.py and beam_interface.py to top level
