+++ src/pyuvdata/uvbeam/__init__.py
@@ -3,6 +3,4 @@
 
 """Init file for UVBeam."""
 
-from .analytic_beam import AiryBeam, GaussianBeam, ShortDipoleBeam, UniformBeam  # noqa
-from .beam_interface import BeamInterface  # noqa
 from .uvbeam import *  # noqa
